+++ src/routes/[...collection]/+page.ts
@@ -1,5 +1,4 @@
 import type { PageLoad } from './$types';
-import { GetCollection, ViewModes } from '$lib/db';
 
 export const load: PageLoad<CollectionView> = async ({ params }) => {
     const [collectionId, collectionName, editModeId, viewModeId] = params.collection.split('/');
@@ -8,6 +7,5 @@ export const load: PageLoad<CollectionView> = async ({ params }) => {
         collectionName: collectionName,
         editModeId: +editModeId,
         viewModeId: +viewModeId,
-        notes: await GetCollection(+collectionId, ViewModes[+viewModeId].type)
     };
 };
